% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_loss.R
\name{compute_loss}
\alias{compute_loss}
\title{Compute the Loss for Two Feature Allocations}
\usage{
compute_loss(Z1, Z2)
}
\arguments{
\item{Z1}{something}

\item{Z2}{something}
}
\value{
The loss as a scalar value.
}
\description{
The loss between two feature allocations is computed.
}
\examples{
Z1 <- matrix(c(0,0,0,0,0,0,0,0,1,0,0,1), byrow=TRUE, nrow=6)
Z2 <- matrix(c(0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,0), byrow=TRUE, nrow=6)
compute_loss(Z1,Z2)

}
