% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{fangs-package}
\alias{fangs-package}
\title{fangs: Feature Allocation Neighborhood Greedy Search Algorithm}
\description{
A neighborhood-based, greedy search algorithm is performed to estimate a feature allocation by minimizing the expected loss based on posterior samples from the feature allocation distribution. The method is described in Dahl, Johnson, and Andros (2023) "Comparison and Bayesian Estimation of Feature Allocations" \doi{10.1080/10618600.2023.2204136}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbdahl/fangs-package}
  \item Report bugs at \url{https://github.com/dbdahl/fangs-package/issues}
}

}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item R. Jacob Andros \email{androsrj@gmail.com} (\href{https://orcid.org/0000-0002-1289-385X}{ORCID})
  \item Devin J. Johnson \email{devin.j.johnson7@gmail.com} (\href{https://orcid.org/0000-0003-2619-6649}{ORCID})
}

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
