% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangs2.R
\name{fangs2}
\alias{fangs2}
\title{Feature Allocation Neighborhood Greedy Search}
\usage{
fangs2(
  Zs,
  maxNFeatures = 0,
  nSamples = length(Zs),
  nBest = 1,
  k = 0,
  probFlip = NULL,
  maxIter = 100,
  nCores = 0
)
}
\arguments{
\item{Zs}{something}

\item{maxNFeatures}{The maximum number of features that can be considered by
the optimization algorithm, which has important implications for the
interpretability of the resulting feature allocation. If the supplied value
is zero, there is no constraint.}

\item{nBest}{something}

\item{k}{something}

\item{probFlip}{something}

\item{maxIter}{something}

\item{nCores}{The number of CPU cores to use, i.e., the number of
simultaneous runs at any given time. A value of zero indicates to use all
cores on the system.}
}
\value{
A list of the following elements:
\itemize{
\item Z - The letters of the alphabet.
\item bestDraw - A vector of numbers.
\item loss - A vector of numbers.
\item best - A vector of numbers.
\item k - A vector of numbers.
\item probFlip - A vector of numbers.
\item iter - A vector of numbers.
}
}
\description{
An implementation of the feature allocation greedy search algorithm is
provided.
}
\examples{
1 + 3

}
