% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{fangs-package}
\alias{fangs-package}
\title{fangs: Feature Allocation Neighborhood Greedy Search Algorithm}
\description{
A neighborhood-based, greedy search algorithm is performed to estimate a feature allocation by minimizing the expected loss based on posterior samples from the feature allocation distribution. The method is described in Dahl, Johnson, and Andros (2023) "Comparison and Bayesian Estimation of Feature Allocations" \doi{10.1080/10618600.2023.2204136}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbdahl/fangs-package}
  \item Report bugs at \url{https://github.com/dbdahl/fangs-package/issues}
}

}
\author{
\strong{Maintainer}: David B. Dahl \email{dahl@stat.byu.edu} (\href{https://orcid.org/0000-0002-8173-1547}{ORCID})

Authors:
\itemize{
  \item R. Jacob Andros \email{androsrj@gmail.com} (\href{https://orcid.org/0000-0002-1289-385X}{ORCID})
  \item Devin J. Johnson \email{devin.j.johnson7@gmail.com} (\href{https://orcid.org/0000-0003-2619-6649}{ORCID})
}

Other contributors:
\itemize{
  \item Alex Crichton \email{alex@alexcrichton.com} (Rust crates: cfg-if, proc-macro2) [contributor]
  \item Andrii Dmytrenko \email{andrii.dmytrenko@deliveroo.co.uk} (Rust crate: lapjv) [contributor]
  \item Brendan Zabarauskas \email{bjzaba@yahoo.com.au} (Rust crate: approx) [contributor]
  \item David B. Dahl \email{dahl@stat.byu.edu} (Rust crates: roxido, roxido_macro) [contributor]
  \item David Tolnay \email{dtolnay@gmail.com} (Rust crates: proc-macro2, quote, syn, unicode-ident) [contributor]
  \item Jim Turner (Rust crate: ndarray) [contributor]
  \item Josh Stone \email{cuviper@gmail.com} (Rust crates: autocfg, rayon, rayon-core) [contributor]
  \item Niko Matsakis \email{niko@alum.mit.edu} (Rust crates: rayon, rayon-core) [contributor]
  \item R. Janis Goldschmidt (Rust crate: matrixmultiply) [contributor]
  \item The Cranelift Project Developers (Rust crate: wasi) [contributor]
  \item The CryptoCorrosion Contributors (Rust crates: ppv-lite86, rand_chacha) [contributor]
  \item The Rand Project Developers (Rust crates: getrandom, rand, rand_chacha, rand_core, rand_pcg) [contributor]
  \item The Rust Project Developers (Rust crates: libc, log, num-complex, num-integer, num-traits, rand, rand_chacha, rand_core) [contributor]
  \item Ulrik Sverdrup "bluss" (Rust crate: ndarray) [contributor]
  \item bluss (Rust crates: either, itertools, matrixmultiply, rawpointer) [contributor]
}

}
\keyword{internal}
