% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangs.R
\name{fangs}
\alias{fangs}
\title{Feature Allocation Neighborhood Greedy Search}
\usage{
fangs(
  samples,
  nIterations = 1000,
  maxNFeatures = 0,
  nCandidates = length(samples),
  nBests = 1,
  nCores = 0,
  quiet = FALSE
)
}
\arguments{
\item{samples}{An object of class 'list' containing posterior samples from a feature allocation distribution in binary matrix form.}

\item{nIterations}{The number of iterations (i.e., proposed changes) to consider per proposed estimate.}

\item{maxNFeatures}{The maximum number of features that can be considered by
the optimization algorithm, which has important implications for the
interpretability of the resulting feature allocation. If the supplied value
is zero, there is no constraint.}

\item{nCandidates}{The number of feature allocations from the provided samples
in \code{Z} for which the expected loss is estimated.}

\item{nBests}{The number of samples among \code{nCandidates} feature allocations which are chosen (by lowest expected loss) to be tuned in the sweetening phase.}

\item{nCores}{The number of CPU cores to use, i.e., the number of
simultaneous runs at any given time. A value of zero indicates to use all
cores on the system.}

\item{quiet}{If \code{TRUE}, intermediate status reporting is suppressed.}
}
\value{
A list of the following elements:
\itemize{
\item estimate - The feature allocation estimate in binary matrix form.
\item loss - The estimated expected loss of the estimate.
\item iteration - The iteration number (out of \code{nIterations}) at which the estimate was found.
\item seconds - The elapsed time of the entire algorithm.
\item whichBest - The proposal number (out of \code{nBests}) from which the estimate was found.
\item nBests - The original supplied value of \code{nBests}.
\item nCandidates - The original supplied value of \code{nCandidates}.
\item maxNFeatures - The original supplied value of \code{maxNFeatures}.
}
}
\description{
An implementation of the feature allocation greedy search algorithm is
provided.
}
\examples{
# To reduce load on CRAN servers, limit the number of iterations, but not necessary in practice.
data(samplesFA)
fangs(samplesFA, nIterations=100)
fangs(samplesFA, nIterations=50, nCandidates=length(samplesFA)/2, nBests=3, quiet=TRUE)

}
