% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangs2.R
\name{fangs2}
\alias{fangs2}
\title{Feature Allocation Neighborhood Greedy Search}
\usage{
fangs2(Zs, nBest = 1, k = 1, prob1 = -1, maxIter)
}
\arguments{
\item{Zs}{something}

\item{nBest}{something}

\item{k}{something}

\item{prob1}{something}

\item{maxIter}{something}
}
\value{
A list of the following elements:
\itemize{
\item Z - The letters of the alphabet.
\item bestDraw - A vector of numbers.
\item loss - A vector of numbers.
\item best - A vector of numbers.
\item k - A vector of numbers.
\item prob1 - A vector of numbers.
\item iter - A vector of numbers.
}
}
\description{
An implementation of the feature allocation greedy search algorithm is
provided.
}
\examples{
1 + 3

}
