% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fangs2.R
\name{fangs2}
\alias{fangs2}
\title{Feature Allocation Neighborhood Greedy Search}
\usage{
fangs2(
  Zs,
  nIterations = 100,
  maxNFeatures = 0,
  nCandidates = length(Zs),
  nBests = 1,
  nCores = 0
)
}
\arguments{
\item{Zs}{something}

\item{nIterations}{The number of iterations (i.e., proposed changes) to consider.}

\item{maxNFeatures}{The maximum number of features that can be considered by
the optimization algorithm, which has important implications for the
interpretability of the resulting feature allocation. If the supplied value
is zero, there is no constraint.}

\item{nCandidates}{The number of feature allocations from the provided samples
in \code{Zs} to use to seed the algorithm.}

\item{nBests}{Among \code{nCandidates} feature allocations, how many should be
sweetened when searching for the best estimate.}

\item{nCores}{The number of CPU cores to use, i.e., the number of
simultaneous runs at any given time. A value of zero indicates to use all
cores on the system.}
}
\value{
A list of the following elements.  \emph{This needs to be updated}:
\itemize{
\item Z - The letters of the alphabet.
\item bestDraw - A vector of numbers.
\item loss - A vector of numbers.
\item best - A vector of numbers.
\item k - A vector of numbers.
\item probFlip - A vector of numbers.
\item iter - A vector of numbers.
}
}
\description{
An implementation of the feature allocation greedy search algorithm is
provided.
}
\examples{
1 + 3

}
